+++ storage/storage.py
@@ -41,9 +41,7 @@ class Storage:
 
     def query(self, query: str):
         with self.engine.connect() as connection:
-            result = connection.execute(
-                text(query)
-            )
+            result = connection.execute(text(query))
         return result.fetchall()
 
     def add_triplets(self, triplets: list[list]) -> None:
@@ -52,46 +50,34 @@ class Storage:
             session.bulk_insert_mappings(Triplets, mappings)
             session.commit()
 
-    def get_references_containing_taxon(self, tid: int) -> set[int]:
+    def get_generic_of_generic(self, out: Any, inp: Any, item: int) -> set[int]:
         with self.session() as session:
-            result = session.query(Triplets.reference_id).filter(Triplets.taxon_id == tid).distinct()
+            result = session.query(out).filter(inp == item).distinct()
             return {row[0] for row in result}
+
+    def get_references_containing_taxon(self, tid: int) -> set[int]:
+        return self.get_generic_of_generic(Triplets.reference_id, Triplets.taxon_id, tid)
 
     def get_references_containing_structure(self, sid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.reference_id).filter(Triplets.structure_id == sid).distinct()
-            return {row[0] for row in result}
-
-    def get_references_of_couple(self, sid: int, tid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.reference_id).filter(Triplets.structure_id == sid, Triplets.taxon_id == tid).distinct()
-            return {row[0] for row in result}
+        return self.get_generic_of_generic(Triplets.reference_id, Triplets.structure_id, sid)
 
     def get_taxa_of_reference(self, rid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.taxon_id).filter(Triplets.reference_id == rid).distinct()
-            return {row[0] for row in result}
+        return self.get_generic_of_generic(Triplets.taxon_id, Triplets.reference_id, rid)
 
     def get_structures_of_reference(self, rid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.structure_id).filter(Triplets.reference_id == rid).distinct()
-            return {row[0] for row in result}
+        return self.get_generic_of_generic(Triplets.structure_id, Triplets.reference_id, rid)
 
     def get_taxa_of_structure(self, sid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.taxon_id).filter(Triplets.structure_id == sid).distinct()
-            return {row[0] for row in result}
+        return self.get_generic_of_generic(Triplets.taxon_id, Triplets.structure_id, sid)
 
     def get_structures_of_taxon(self, tid: int) -> set[int]:
-        with self.session() as session:
-            result = session.query(Triplets.structure_id).filter(Triplets.taxon_id == tid).distinct()
-            return {row[0] for row in result}
+        return self.get_generic_of_generic(Triplets.structure_id, Triplets.taxon_id, tid)
 
     def validate_column_name(self, s):
         ps = s.replace('_', '')
         return ps.islower() and ps.isalpha()
 
-    def get_generics_of_generics(self, inp: Any, out: Any, items: set[int]) -> set[int]:
+    def get_generics_of_generics(self, out: Any, inp: Any, items: set[int]) -> set[int]:
         if len(items) > self.list_limit:
             raise Exception(
                 f"Too many {inp}, the limit is {self.list_limit} and you have {len(items)}"
@@ -102,16 +88,37 @@ class Storage:
             return {row[0] for row in result}
 
     def get_references_of_structures(self, structures: set[int]) -> set[int]:
-        return self.get_generics_of_generics(Triplets.structure_id, Triplets.reference_id, structures)
+        return self.get_generics_of_generics(Triplets.reference_id, Triplets.structure_id, structures)
 
     def get_references_of_taxa(self, taxa: set[int]) -> set[int]:
-        return self.get_generics_of_generics(Triplets.taxon_id, Triplets.reference_id, taxa)
+        return self.get_generics_of_generics(Triplets.reference_id, Triplets.taxon_id, taxa)
 
     def get_structures_of_references(self, references: set[int]) -> set[int]:
-        return self.get_generics_of_generics(Triplets.reference_id, Triplets.structure_id, references)
+        return self.get_generics_of_generics(Triplets.structure_id, Triplets.reference_id, references)
 
     def get_taxa_of_structures(self, structures: set[int]) -> set[int]:
-        return self.get_generics_of_generics(Triplets.structure_id, Triplets.taxon_id, structures)
+        return self.get_generics_of_generics(Triplets.taxon_id, Triplets.structure_id, structures)
 
     def get_taxa_of_references(self, references: set[int]) -> set[int]:
-        return self.get_generics_of_generics(Triplets.reference_id, Triplets.taxon_id, references)
+        return self.get_generics_of_generics(Triplets.taxon_id, Triplets.reference_id, references)
+
+    def get_triplets_for(self,
+                         reference_ids: set[int] | None,
+                         structure_ids: set[int] | None,
+                         taxon_ids: set[int] | None) -> set[tuple[int, int, int]]:
+        with self.session() as session:
+            filters = []
+            if reference_ids is not None:
+                filters += [Triplets.reference_id.in_(reference_ids)]
+            if structure_ids is not None:
+                filters += [Triplets.structure_id.in_(structure_ids)]
+            if taxon_ids is not None:
+                filters += [Triplets.taxon_id.in_(taxon_ids)]
+            result = session.query(Triplets.reference_id,
+                                   Triplets.structure_id,
+                                   Triplets.taxon_id)
+
+            if len(filters) > 0:
+                result = result.filter(*filters)
+
+            return {(row[0], row[1], row[2]) for row in result.distinct()}